fix parser import cache bug
When optimistically trying to import the parser, if import fails
it will also fail even after parser is built. To fix this required
reloading vscode, which is not a great UX. A better approach here
is to check if the `prebuilds` dir exists, and only build the parser
if it does not.